Improved test coverage and renaming
