+++ tests/test-cases/writers/skoscodelistwriter/v1.0/colourconfig.json
