+++ tests/test-cases/writers/skoscodelistwriter/v1.0/colours.csv
